Pull out non-rendering logic into separate hooks for notebook cells
